proc: Don't let Google Photos run in background on battery
Google Photos goes insane for some reason with nonstop CPU usage of
100%, absolutely obliterating battery life. Kill it when it's in the
background and the battery is discharging, otherwise let it live when
the battery is charging so that it can back up photos. It'll start
itself automatically when the charger is connected so there's no need to
worry about that.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>